refactor: Change structure to fit new crate layout.
