+++ src/rust/ide/view/graph-editor/src/component/node/icon.rs
