+++ src/rust/ide/view/graph-editor/src/component/text_list.rs
@@ -6,8 +6,6 @@ use crate::prelude::*;
 use enso_frp as frp;
 use enso_frp;
 use ensogl::data::color;
-use ensogl::display::Attribute;
-use ensogl::display::Buffer;
 use ensogl::display::Sprite;
 use ensogl::display::scene::Scene;
 use ensogl::display::shape::*;
